+++ openai_realtime_voice_agent/root/run.sh
@@ -15,6 +15,7 @@ MAX_OUTPUT_TOKENS=$(bashio::config 'max_output_tokens')
 # --- 💬 Conversation ---
 FOLLOW_UP_LISTEN_SECONDS=$(bashio::config 'follow_up_listen_seconds')
 FOLLOW_UP_OPEN_DELAY_MS=$(bashio::config 'follow_up_open_delay_ms')
+WAKE_OPEN_DELAY_MS=$(bashio::config 'wake_open_delay_ms')
 VAD_EAGERNESS=$(bashio::config 'vad_eagerness')
 PHASE_IDLE_DEBOUNCE_MS=$(bashio::config 'phase_idle_debounce_ms')
 
@@ -56,6 +57,7 @@ export OPENAI_SPEED
 export MAX_OUTPUT_TOKENS
 export FOLLOW_UP_LISTEN_SECONDS
 export FOLLOW_UP_OPEN_DELAY_MS
+export WAKE_OPEN_DELAY_MS
 export VAD_EAGERNESS
 export PHASE_IDLE_DEBOUNCE_MS
 export ENABLE_WEB_SEARCH